refactor(gateway): extract the Gateway's Status Enum to the health package
